requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-q>=1.3
Pillow>=10.0
numpy>=1.24
opencv-python-headless>=4.8  
drf-yasg>=1.21              
djangorestframework-simplejwt>=5.3
torch 
torchvision 
torchaudio
dtlpy
setuptools>=78.1.1 
python-dotenv
django-crum
pandas
openpyxl
google-genai